pyproject.toml
[project]
name = "restate-webserice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.5",
    "httpx>=0.27.2",
    "hypercorn>=0.17.3",
    "restate-sdk>=0.4.1",
    "sqlmodel>=0.0.22",
]
